Delete with ajax, add meta tag with csrf token which auto generate thymeleaf.

+++ src/main/java/com/fpt/controller/admin/AuthorController.java
@@ -5,10 +5,13 @@ import com.fpt.entity.Author;
 import com.fpt.service.admin.AuthorServiceImpl;
 import com.fpt.specification.AuthorSpecification;
 import com.fpt.specification.SearchCriteria;
+import com.google.gson.Gson;
+import com.sun.deploy.net.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -17,6 +20,7 @@ import org.springframework.util.StringUtils;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 @Controller
@@ -81,11 +85,28 @@ public class AuthorController {
         return "admin/author/list";
     }
 
-    @RequestMapping(method = RequestMethod.POST, value = "/delete/{id}")
+    @RequestMapping(method = RequestMethod.GET, value = "/delete/{id}")
     public String delete(@PathVariable(value = "id", required = false) long id, RedirectAttributes redirectAttributes) {
+        Author author = authorService.getById(id);
+        if (author == null) {
+            return "error/404";
+        }
         authorService.delete(id);
         redirectAttributes.addFlashAttribute("Success!", "Deleted contact successfully!");
         return "redirect:" + ProjectConfig.PREFIX_ADMIN + ProjectConfig.PREFIX_ADMIN_AUTHORS;
+    }
+
+    @RequestMapping(method = RequestMethod.DELETE, value = "/ajax-delete/{id}")
+    @ResponseBody
+    public String deleteWithAjax(@PathVariable(value = "id", required = false) long id, HttpServletResponse response) {
+        Author author = authorService.getById(id);
+        if (author == null) {
+            response.setStatus(HttpStatus.NOT_FOUND.value());
+            return new Gson().toJson("Error");
+        }
+        authorService.delete(id);
+        response.setStatus(HttpStatus.OK.value());
+        return new Gson().toJson("Ok");
     }
 
     @RequestMapping(method = RequestMethod.GET, value = "/update/{id}")
